pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flair-benchmark"
version = "1.0.0"
description = "Federated Learning Assessment for ICU Research - Privacy-first benchmark for CLIF data"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "CLIF Consortium", email = "clif_consortium@uchicago.edu" }
]
keywords = [
    "healthcare",
    "icu",
    "benchmark",
    "federated-learning",
    "clif",
    "machine-learning",
    "clinical-research"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "polars>=0.20.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "clifpy>=0.3.4",
    "reportlab>=4.4.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://clif-icu.com"
Repository = "https://github.com/clif-consortium/flair"
Documentation = "https://clif-icu.com/flair"

[tool.hatch.build.targets.wheel]
packages = ["flair_benchmark"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
